cleanup for evaluations
