+++ components/proposals/ProposalPage/components/ProposalEvaluations/components/Review/components/StepperIcon.tsx
