+++ components/proposals/ProposalPage/components/ProposalEvaluations/components/Review/components/VoteEvaluation/components/SnapshotVoteDetails.tsx
