+++ components/proposals/ProposalPage/components/ProposalEvaluations/components/Review/components/VoteEvaluation/components/SnapshotVotingForm/components/WeightedVoting.tsx
