+++ lib/proposal/__tests__/createProposal.spec.ts
@@ -324,6 +324,8 @@ describe('Creates a page and proposal with relevant configuration', () => {
       spaceId: space.id,
       userId: user.id,
       workflowId: proposalWorkflow.id,
+      authors: [user.id],
+      pageProps: { title: 'test' },
       evaluations: evaluationTemplate.map((item, index) => ({
         id: item.id,
         rubricCriteria: [],
